ReversiHumanPlayer
I did the ReversiHumanPlayer. I think that all my group members and I did well on this program. In addition, we spent a lot of hours coding it. Based on our efforts, we should full credits, but I am sured that this program needs some improvements. For example, if I have more time on this project, we can make this  board larger, and people will enjoy this game more than before.